TrainingBlock Create and Get
Started the overall structure of the back end API. Created entities mappers and overall structure that this backend system will use.

Overall first apis that need to get working are creating training blocks. Then need to move on to weekds days and excercise.

Overall Idea is that a training block is created and within the training block weeks and days and excercise can be configured and added.